requirements.txt
mcp
mcp[cli]
httpx
pydantic
python-dotenv
python-dateutil
structlog
pytest
pytest-asyncio
pytest-httpx
black
requests
aiohttp
uuid